requirements.txt
Flask==2.2.3
requests==2.28.1
Werkzeug==2.2.3
gunicorn==21.2.0
pycairo
PyGObject 